Refactor dataflow classes (#31389)
This change simplifies our data flow analysis by removing an obsolete subclass and renaming some of the components to standardized names used in the literature on lattice data flow analysis.

Fixes #23207
