+++ src/Compilers/CSharp/Portable/FlowAnalysis/AbstractRegionDataFlowPass.cs
@@ -10,7 +10,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 {
     // Note: this code has a copy-and-paste sibling in AbstractRegionDataFlowPass.
     // Any fix to one should be applied to the other.
-    internal class AbstractRegionDataFlowPass : DataFlowPass
+    internal class AbstractRegionDataFlowPass : DefiniteAssignmentPass
     {
         internal AbstractRegionDataFlowPass(
             CSharpCompilation compilation,
@@ -30,7 +30,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         /// </summary>
         protected override ImmutableArray<PendingBranch> Scan(ref bool badRegion)
         {
-            SetState(ReachableState());
+            SetState(TopState());
             MakeSlots(MethodParameters);
             if ((object)MethodThisParameter != null) GetOrCreateSlot(MethodThisParameter);
             var result = base.Scan(ref badRegion);
