+++ src/Compilers/CSharp/Portable/FlowAnalysis/UnassignedAddressTakenVariablesWalker.cs
@@ -12,7 +12,7 @@ namespace Microsoft.CodeAnalysis.CSharp
     /// An analysis that computes the set of variables that may be used
     /// before being assigned anywhere within a method.
     /// </summary>
-    internal class UnassignedAddressTakenVariablesWalker : DataFlowPass
+    internal class UnassignedAddressTakenVariablesWalker : DefiniteAssignmentPass
     {
         private UnassignedAddressTakenVariablesWalker(CSharpCompilation compilation, Symbol member, BoundNode node)
             : base(compilation, member, node)
