+++ src/Compilers/CSharp/Portable/FlowAnalysis/UnassignedVariablesWalker.cs
@@ -12,7 +12,7 @@ namespace Microsoft.CodeAnalysis.CSharp
     /// An analysis that computes the set of variables that may be used
     /// before being assigned anywhere within a method.
     /// </summary>
-    internal class UnassignedVariablesWalker : DataFlowPass
+    internal class UnassignedVariablesWalker : DefiniteAssignmentPass
     {
         private UnassignedVariablesWalker(CSharpCompilation compilation, Symbol member, BoundNode node)
             : base(compilation, member, node, new NeverEmptyStructTypeCache())
